+++ AdditionalNetworking/src/Patches/StartOfRoundPatch.cs
@@ -32,6 +32,10 @@ internal class StartOfRoundPatch
             PlayerNetworking.Instance.ValidClientIDs.Remove(clientId);
         if (ShotgunNetworking.Instance != null)
             ShotgunNetworking.Instance.ValidClientIDs.Remove(clientId);
+        if (BoomboxNetworking.Instance != null)
+            BoomboxNetworking.Instance.ValidClientIDs.Remove(clientId);
+        if (NutcrackerNetworking.Instance != null)
+            NutcrackerNetworking.Instance.ValidClientIDs.Remove(clientId);
     }
     
     [HarmonyFinalizer]