+++ CHANGELOG.md
@@ -1,2 +1,9 @@
-### v1.0.0
-- WIP
+### v1.0.1
+- Add Enemy transform sync
+- Add Nutcracker torso rotation sync
+- only compute Nutcracker rotation on Owner
+
+### v1.0.1
+- Add Item transform sync
+- only set item position to floor once
+- only perform fall calculations on Owner